- Fixed whitespace issues detected by latest stylecop update - Added additional tests for serialization
