+++ src/MooVC.Tests/Processing/ThreadSafeHostedServiceTests/WhenStopAsyncIsCalled.cs
