+++ src/MooVC.Tests/Processing/ThreadSafeHostedServiceTests/WhenTryStartAsyncIsCalled.cs
