+++ src/MooVC.Tests/Processing/ThreadSafeHostedServiceTests/WhenTryStopAsyncIsCalled.cs
