+++ src/MooVC/Collections/Generic/EnumerableExtensions.ForAll.cs
