+++ src/MooVC/Serialization/Resources.resx
@@ -117,6 +117,9 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="SerializationInfoExtensionsCreateMethodInfoFailure" xml:space="preserve">
+    <value>The required deserialization method is not supported by the current version of the .NET framework.</value>
+  </data>
   <data name="SerializationInfoExtensionsGetValueTypeInvalid" xml:space="preserve">
     <value>Element {0} could not be deserialized to type {1}.</value>
   </data>
