+++ src/MooVC/Serialization/SerializationInfoExtensions.AddInternalValue.cs
